feat: update default date values in salary and payroll reports
Updated the default values for the 'from_date' and 'to_date' fields in the Employee Salary Register, Kenya Bank Payroll Advice, and Kenya Payroll Register reports to use the new date functions. The 'from_date' now defaults to two months prior to the current month's start, and the 'to_date' defaults to two months prior to the current month's end, enhancing the accuracy of date selections in these reports.